Add color and shape-based genome (#1)
When a new cell is born it will take on the color and shape that is most commonly represented by its parents. If there is not a most common trait expressed, take a random trait from a parent.

Every birth is accompanied by a 10% chance of mutation, which will select a completely random color and shape. This has the effect of making fliers and oscillators change color and shape over time.